Redirect subcommands I/O to the parent process, fixing blocking issues
By default subcommands' stderr, stdout and stdin are redirected
to pipes accessible from Process.getErrorStream(),
Process.getOutputStream() and Process.getInputStream().

This means in particular that, when subcommands write to stderr and stdout,
the writes will be blocking until someone reads from
Process.getErrorStream() and Process.getOutputStream().

Since we don't, the subcommands used to block as soon as they needed to
write to stderr or stdout.

I suspect the issue was not obvious because writes to stderr and stdout
in Elasticsearch are done through Apache Commons Logs, and those writes
are generally buffered, meaning they won't block until the buffer is
filled.