Issue 105: add support for inter process synchronization
Multiple plugins could run in parallel and only one of them will run in
master mode, the rest will run in slave mode. The master plugin is
responsible for downloading and installing the ES artifact in the local
repo, while the slave plugins will wait for the master to finish the
process in order to resolve the artifact. All this so that the plugin
makes efficient use of bandwidth, in order to avoid downloading the same
artifact multiple times.