Fix ForkedElasticsearchProcessDestroyer issues
On our Jenkins server we regularly see the following message when the
plugin tries to start the elasticsearch server:
  "This Elasticsearch process destroyer does not support this operation"
This started to happen very frequently after I upgraded to version 6.17
of the plugin. I've seen several similar issues reported, but they're all
closed as resolved. However, since we're still getting this error I had a
look at the code of the ForkedElasticsearchProcessDestroyer.
I do see some issues:
1. The shutdown hook registered in the constructor can throw a
   NullPointerException, because the process can be null.
2. If the DefaultExecutor calls add on the process destroyer then it will
   also call remove on the process destroyer because remove is called
   in a finally block. Based on the stack trace from the Jenkins logging
   I think this is the root cause of my problem.

So the contract for the implementor of ProcessDestroyer seems to be:
   for every process added a call to remove it again will always follow.
Unfortunately this is not very clear from javadoc.  or project documentation.
The solution I chose:
- only throw an IllegalStateException when the process to be removed is not
  equal to the one that was added. This fixes the root cause.
- Let ForkedElasticsearchProcessDestroyer implement Runnable and only call
  terminateProcess() if process is not null. This will prevent the NPEs.

The reason do not I register the shutdown hook in the constructor is to make it
easier to test the implementation of ForkedElasticsearchProcessDestroyer.